+++ lib/griddle.dart
@@ -1,5 +1,4 @@
 import 'dart:io' as io;
-import 'dart:math' as math;
 
 import 'package:meta/meta.dart';
 import 'package:neoansi/neoansi.dart';